test steps rework

+++ TestSteps/include/TestSteps/TestSteps.h
@@ -57,9 +57,8 @@ class TestSteps
 {
 public:
     typedef unsigned TestStepT;
-    typedef unsigned short TestLineT;
+    typedef unsigned TestLineT;
     typedef unsigned char StepLevelT;
-
 
     //  clear test steps
     static void clear();
